Changed data type for parameter avg from int to double to accurately portray the average (as it will now include decimals) in calculation.hpp, output.hpp, and main.cpp. In calculation.hpp, parameters a and b also had to be changed from int to double to divide and produce the proper output.

+++ output.hpp
@@ -1,6 +1,6 @@
 #include <iostream>
 using namespace std;
 
-void output(int sum, int avg){
+void output(int sum, double avg){
      cout << "Sum: " << sum << " Average: " << avg << endl;
 }
